requirements.txt
langchain
langchain-core
langchain-openai
tavily-python
streamlit
langchain-community
yfinance
langchain-experimental
tabulate